Set up initial email functionality
Got set up with SendGrid to provide SMTP service for sending actual
emails. Right now it's just sending emails to KeystoneJS admins when
there's a new enquiry.

The emails need styled, and they're set up in an over-complicated
manner. They need to be simplified and given NGC branding. We'll
eventually re-use this for sending emails to users, etc.

Nodemailer's integration with SendGrid is a bit sketchy right now: we're
relying on a dumb fork of the no-longer-officially supported library.

See this issue:
https://github.com/sendgrid/nodemailer-sendgrid-transport/issues/29

The actual SendGrid account is managed through Azure.